Add some method to encrypt and decrypt
Simple ECB encryption using a stream, it works with input.txt
but if fails when the file gets larger. The error it throws is:
BadPaddingException: Given final block not properly padded
All the answers I have read talk about a problem with the
private key (Which I doubt to be the case here)

I though that the bug could come from the way fs2 makes chunks,
e.g if there are many chunks that need padding during the
encryptin (because the chunk size is not multiple of the block
size), maybe this could cause the decrypt algorith to fail (because
the padding is expected only at the end of the plaintext). So
I tried to separate the message I wanted to encrypt into pieces
(which size is not a multiple of 8) encrypt them separatly with
padding, then combine them and decrypt them as one single input,
and it works fine without using streams. So it should be able to
decrypt the message by chunks too.

I though it failed with bigger files because there was for sure
several chunks (and that it worked with smaller files because there
was a single chunk), but it works in examples that use several
chunks too, so I don't understand yet why.