Add EncryptedClient, push and get work
Weird bug: when you push a file to the server and then you get it
back, the server adds 3 bytes to the file which make the decryption
fail. Even without decrypting the message in the client side the
file is corrupted (3 extra bytes).

This does not happen if the server is reset. so if you up the server
then push a file, down the server, up the server, get the file,
everything works fine.