Fix streaming issues,
The bug was cause by the fact that the server was sending "OK"
and just afterwards the Stream[F, Byte] of the file, so part
of the stream was being comsumed by the client while reading
"OK". Something that misleaded me for finding the bug was that
the bug was only present after the first request, (e.g if you up
the server and send a PUSH or GET request, it worked fine, but
the next one failed), so I thought the bug was comming from
the way the server deals with sockets.
The solution I have found (surely not perfect), is that the part
that send the stream of data is the one that listen for the last
"OK" message.
Before:
Client  -GET-> Server
Client <--OK-  Server
Client <-Data- Server     Here the data get mixed with "OK"

Now:
Client  -GET-> Server
Client <--OK-  Server
Client  --OK-> Server
Client <-Data- Server

Refactor of the client part so I can create several clients
easily, (e.g DebugClient)

Add Some debug methods like showChunks